Change package in pubnub-spaces module: entities -> space
