--- pubnub-memberships/src/main/kotlin/com/pubnub/entities/models/consumer/membership/SpaceDetailsLevel.kt
@@ -1,15 +1,0 @@
-package com.pubnub.entities.models.consumer.membership
-
-import com.pubnub.api.models.consumer.objects.membership.PNChannelDetailsLevel
-
-enum class SpaceDetailsLevel {
-    SPACE,
-    SPACE_WITH_CUSTOM
-}
-
-internal fun SpaceDetailsLevel.toPNChannelDetailsLevel(): PNChannelDetailsLevel {
-    return when (this) {
-        SpaceDetailsLevel.SPACE -> PNChannelDetailsLevel.CHANNEL
-        SpaceDetailsLevel.SPACE_WITH_CUSTOM -> PNChannelDetailsLevel.CHANNEL_WITH_CUSTOM
-    }
-}
