--- pubnub-memberships/src/main/kotlin/com/pubnub/entities/models/consumer/membership/UserMembershipsResultKey.kt
@@ -1,17 +1,0 @@
-package com.pubnub.entities.models.consumer.membership
-
-import com.pubnub.api.models.consumer.objects.PNMembershipKey
-import com.pubnub.api.models.consumer.objects.SortableKey
-
-enum class UserMembershipsResultKey : SortableKey {
-    SPACE_ID, SPACE_NAME, SPACE_UPDATED, UPDATED;
-
-    fun toPNMembershipKey(): PNMembershipKey {
-        return when (this) {
-            SPACE_ID -> PNMembershipKey.CHANNEL_ID
-            SPACE_NAME -> PNMembershipKey.CHANNEL_NAME
-            SPACE_UPDATED -> PNMembershipKey.CHANNEL_UPDATED
-            UPDATED -> PNMembershipKey.UPDATED
-        }
-    }
-}
