+++ pubnub-memberships/src/main/kotlin/com/pubnub/space/models/consumer/membership/SpaceDetailsLevel.kt
@@ -1,0 +1,15 @@
+package com.pubnub.space.models.consumer.membership
+
+import com.pubnub.api.models.consumer.objects.membership.PNChannelDetailsLevel
+
+enum class SpaceDetailsLevel {
+    SPACE,
+    SPACE_WITH_CUSTOM
+}
+
+internal fun SpaceDetailsLevel.toPNChannelDetailsLevel(): PNChannelDetailsLevel {
+    return when (this) {
+        SpaceDetailsLevel.SPACE -> PNChannelDetailsLevel.CHANNEL
+        SpaceDetailsLevel.SPACE_WITH_CUSTOM -> PNChannelDetailsLevel.CHANNEL_WITH_CUSTOM
+    }
+}
