+++ pubnub-memberships/src/main/kotlin/com/pubnub/space/models/consumer/membership/SpaceMembershipResultKey.kt
@@ -1,0 +1,17 @@
+package com.pubnub.space.models.consumer.membership
+
+import com.pubnub.api.models.consumer.objects.PNMemberKey
+import com.pubnub.api.models.consumer.objects.SortableKey
+
+enum class SpaceMembershipResultKey : SortableKey {
+    USER_ID, USER_NAME, USER_UPDATED, UPDATED;
+
+    fun toPNMemberKey(): PNMemberKey {
+        return when (this) {
+            USER_ID -> PNMemberKey.UUID_ID
+            USER_NAME -> PNMemberKey.UUID_NAME
+            USER_UPDATED -> PNMemberKey.UUID_UPDATED
+            UPDATED -> PNMemberKey.UPDATED
+        }
+    }
+}
